Added a new level 6 Fixed the timer and death count overflows

+++ include/game.h
@@ -43,7 +43,7 @@ struct Game
     uint16_t gameFrame;
     uint16_t levelFrame;
     uint8_t currentLevel;
-    uint8_t deaths;
+    uint16_t deaths;
     level_t level_data;
     uint16_t mapPixelW;
     uint16_t mapPixelH;
